+++ src/components/global/Modules/Services/Services.css
@@ -1,0 +1,65 @@
+.Content {
+    text-align: left;
+  }
+  
+  #help-text{
+    text-align: center;
+  }
+  
+  #card_title_color{
+    color: #fff;
+  }
+  
+  .tg {
+    border-collapse: collapse;
+    border-spacing: 0;
+    border-color: #ccc;
+    width: 500px;
+  }
+  
+  .tg td {
+    font-family: Arial, sans-serif;
+    font-size: 14px;
+    padding: 10px 5px;
+    border-style: solid;
+    border-width: 1px;
+    overflow: hidden;
+    word-break: normal;
+    border-color: #ccc;
+    color: #333;
+    background-color: #fff;
+  }
+  
+  .tg th {
+    font-family: Arial, sans-serif;
+    font-size: 14px;
+    font-weight: normal;
+    padding: 10px 5px;
+    border-style: solid;
+    border-width: 1px;
+    overflow: hidden;
+    word-break: normal;
+    border-color: #ccc;
+    color: #333;
+    background-color: #f0f0f0;
+  }
+  
+  .tg .tg-dx8v {
+    font-size: 12px;
+    vertical-align: top;
+    text-align: left;
+  }
+  
+  .tableContainer {
+    background-color: aliceblue;
+    padding: 0 25%;
+  }
+  
+  .card{
+    max-width: 345;
+  }
+  
+  .media{
+    height: 0;
+    padding-top: 56.25%;
+  }
